requirements.txt
requests
bs4
pyyaml
mattermost